fix: ER-145: Removed root motion running animation stutter
Not the most elegant fix but it works for now, removing/streamlining the root motion keyframes in the running animation seems to have helped a bit but the stutter still remains - so root motion has been removed from the running state for the foreseeable future.

Only direct example of this happening before that I could find is this thread (https://godotforums.org/discussion/26843/animation-stutter), and even in that Calinou (main Godot contributor) has no solution other than just disabling root motion for the animation.